requirements.txt
streamlit
pandas
scikit-learn
python-dotenv
google-generativeai
